+++ src/components/Event.js
