+++ src/components/Friends.js
